Teams_Bot/HR_Bot/requirements.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
ase==3.25.0
attrs==25.3.0
azure-core==1.35.0
azure-identity==1.23.0
babel==2.17.0
banks==2.1.2
beautifulsoup4==4.13.4
boolean.py==5.0
boto3==1.38.43
botocore==1.38.43
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cryptography==45.0.5
cssselect2==0.8.0
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
dill==0.4.0
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
docling==2.41.0
docling-core==2.43.0
docling-ibm-models==3.8.1
docling-parse==4.1.0
docxcompose==1.4.0
docxtpl==0.20.1
easyocr==1.7.2
elastic==5.2.5.3
elastic-transport==8.17.1
elasticsearch
et_xmlfile==2.0.0
etelemetry==0.3.1
filelock==3.18.0
filetype==1.2.0
fitz
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
greenlet==3.2.3
griffe==1.7.3
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain-core==0.3.69
langchain-text-splitters==0.3.8
langsmith==0.4.6
latex2mathml==3.78.0
lazy_loader==0.4
llama-cloud==0.1.28
llama-cloud-services==0.6.36
llama-index-core==0.12.43
llama-index-instrumentation==0.2.0
llama-index-workflows==0.2.2
llama-parse==0.6.36
looseversion==1.3.0
lxml==5.4.0
markdown-it-py==3.0.0
marko==2.1.4
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
multidict==6.5.1
multiprocess==0.70.18
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
ninja==1.11.1.4
nipype==1.10.0
nltk==3.9.1
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.91.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.11.0
packaging==25.0
pandas==2.3.0
pathlib==1.0.1
pdf2image==1.17.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.2
prov==2.1.1
puremagic==1.29
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydot==4.0.1
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.10.1
pylatexenc==2.10
pymongo==4.13.2
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdfium2==4.30.1
pyphen==0.17.2
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-pptx==1.0.2
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rtree==1.4.0
s3transfer==0.13.0
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.16.0
semchunk==2.2.2
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
spglib==2.6.0
SQLAlchemy==2.0.41
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tifffile==2025.6.11
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
traits==7.0.2
transformers==4.53.2
triton==3.3.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
weasyprint==65.1
webencodings==0.5.1
wheel==0.45.1
wrapt==1.17.2
xlsxwriter==3.2.5
yarl==1.20.1
zopfli==0.2.3.post1
zstandard==0.23.0